create charts in tabs
