+++ src/pages/ratings/number-of-actors.tsx
@@ -1,0 +1,3 @@
+export const NumberOfActors = () => {
+  return <div>Number of Actors</div>;
+};
