+++ src/pages/ratings/top-actors.tsx
@@ -1,0 +1,3 @@
+export const TopActors = () => {
+  return <div> Top Actors</div>;
+};
